+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Bugfix
 * fix missing examples for processor decoder
+* fix calculator silently failing on syntax errors
 
 
 ## 18.1.0